+++ pelican/generators.py
@@ -5,6 +5,7 @@ import random
 import logging
 import datetime
 import subprocess
+import shutil
 
 from codecs import open
 from collections import defaultdict
@@ -12,12 +13,13 @@ from functools import partial
 from itertools import chain
 from operator import attrgetter, itemgetter
 
-from jinja2 import Environment, FileSystemLoader, PrefixLoader, ChoiceLoader
-from jinja2.exceptions import TemplateNotFound
+from jinja2 import (Environment, FileSystemLoader, PrefixLoader, ChoiceLoader,
+                    BaseLoader, TemplateNotFound)
 
-from pelican.contents import Article, Page, Category, is_valid_content
+from pelican.contents import Article, Page, Category, StaticContent, \
+        is_valid_content
 from pelican.readers import read_file
-from pelican.utils import copy, process_translations
+from pelican.utils import copy, process_translations, mkdir_p
 from pelican import signals
 
 
@@ -42,12 +44,12 @@ class Generator(object):
                 os.path.join(self.theme, 'templates')))
         self._templates_path += self.settings.get('EXTRA_TEMPLATES_PATHS', [])
 
-
         theme_path = os.path.dirname(os.path.abspath(__file__))
 
         simple_loader = FileSystemLoader(os.path.join(theme_path,
                                          "themes", "simple", "templates"))
         self.env = Environment(
+            trim_blocks=True,
             loader=ChoiceLoader([
                 FileSystemLoader(self._templates_path),
                 simple_loader,  # implicit inheritance
@@ -61,6 +63,8 @@ class Generator(object):
         # get custom Jinja filters from user settings
         custom_filters = self.settings.get('JINJA_FILTERS', {})
         self.env.filters.update(custom_filters)
+
+        signals.generator_init.send(self)
 
     def get_template(self, name):
         """Return the template by name.
@@ -80,8 +84,10 @@ class Generator(object):
 
         :param path: the path to search the file on
         :param exclude: the list of path to exclude
+        :param extensions: the list of allowed extensions (if False, all
+            extensions are allowed)
         """
-        if not extensions:
+        if extensions is None:
             extensions = self.markup
 
         files = []
@@ -95,9 +101,16 @@ class Generator(object):
             for e in exclude:
                 if e in dirs:
                     dirs.remove(e)
-            files.extend([os.sep.join((root, f)) for f in temp_files
-                if True in [f.endswith(ext) for ext in extensions]])
+            for f in temp_files:
+                if extensions is False or \
+                        (True in [f.endswith(ext) for ext in extensions]):
+                    files.append(os.sep.join((root, f)))
         return files
+
+    def add_filename(self, content):
+        location = os.path.relpath(os.path.abspath(content.filename),
+                                   os.path.abspath(self.path))
+        self.context['filenames'][location] = content
 
     def _update_context(self, items):
         """Update the context with the given items from the currrent
@@ -108,6 +121,35 @@ class Generator(object):
             if hasattr(value, 'items'):
                 value = value.items()
             self.context[item] = value
+
+
+class _FileLoader(BaseLoader):
+
+    def __init__(self, path, basedir):
+        self.path = path
+        self.fullpath = os.path.join(basedir, path)
+
+    def get_source(self, environment, template):
+        if template != self.path or not os.path.exists(self.fullpath):
+            raise TemplateNotFound(template)
+        mtime = os.path.getmtime(self.fullpath)
+        with file(self.fullpath) as f:
+            source = f.read().decode('utf-8')
+        return source, self.fullpath, \
+                lambda: mtime == os.path.getmtime(self.fullpath)
+
+
+class TemplatePagesGenerator(Generator):
+
+    def generate_output(self, writer):
+        for source, dest in self.settings['TEMPLATE_PAGES'].items():
+            self.env.loader.loaders.insert(0, _FileLoader(source, self.path))
+            try:
+                template = self.env.get_template(source)
+                rurls = self.settings.get('RELATIVE_URLS')
+                writer.write_file(dest, template, self.context, rurls)
+            finally:
+                del self.env.loader.loaders[0]
 
 
 class ArticlesGenerator(Generator):
@@ -128,13 +170,6 @@ class ArticlesGenerator(Generator):
 
     def generate_feeds(self, writer):
         """Generate the feeds from the current context, and output files."""
-        if self.settings.get('FEED_ATOM') is None \
-                and self.settings.get('FEED_RSS') is None:
-            return
-        elif self.settings.get('SITEURL') is '':
-            logger.warning(
-                'Feeds generated without SITEURL set properly may not be valid'
-            )
 
         if self.settings.get('FEED_ATOM'):
             writer.write_feed(self.articles, self.context,
@@ -143,6 +178,21 @@ class ArticlesGenerator(Generator):
         if self.settings.get('FEED_RSS'):
             writer.write_feed(self.articles, self.context,
                               self.settings['FEED_RSS'], feed_type='rss')
+
+        if self.settings.get('FEED_ALL_ATOM') or \
+                self.settings.get('FEED_ALL_RSS'):
+            all_articles = list(self.articles)
+            for article in self.articles:
+                all_articles.extend(article.translations)
+            all_articles.sort(key=attrgetter('date'), reverse=True)
+
+            if self.settings.get('FEED_ALL_ATOM'):
+                writer.write_feed(all_articles, self.context,
+                                  self.settings['FEED_ALL_ATOM'])
+
+            if self.settings.get('FEED_ALL_RSS'):
+                writer.write_feed(all_articles, self.context,
+                                  self.settings['FEED_ALL_RSS'], feed_type='rss')
 
         for cat, arts in self.categories:
             arts.sort(key=attrgetter('date'), reverse=True)
@@ -254,7 +304,6 @@ class ArticlesGenerator(Generator):
         self.generate_articles(write)
         self.generate_direct_templates(write)
 
-
         # and subfolders after that
         self.generate_tags(write)
         self.generate_categories(write)
@@ -262,7 +311,7 @@ class ArticlesGenerator(Generator):
         self.generate_drafts(write)
 
     def generate_context(self):
-        """change the context"""
+        """Add the articles into the shared context"""
 
         article_path = os.path.normpath(  # we have to remove trailing slashes
             os.path.join(self.path, self.settings['ARTICLE_DIR'])
@@ -272,6 +321,7 @@ class ArticlesGenerator(Generator):
                 article_path,
                 exclude=self.settings['ARTICLE_EXCLUDES']):
             try:
+                signals.article_generate_preread.send(self)
                 content, metadata = read_file(f, settings=self.settings)
             except Exception, e:
                 logger.warning(u'Could not process %s\n%s' % (f, str(e)))
@@ -280,16 +330,19 @@ class ArticlesGenerator(Generator):
             # if no category is set, use the name of the path as a category
             if 'category' not in metadata:
 
-                if os.path.dirname(f) == article_path:  # if the article is not in a subdirectory
-                    category = self.settings['DEFAULT_CATEGORY']
-                else:
+                if (self.settings['USE_FOLDER_AS_CATEGORY']
+                    and os.path.dirname(f) != article_path):
+                    # if the article is in a subdirectory
                     category = os.path.basename(os.path.dirname(f))\
-                                .decode('utf-8')
+                        .decode('utf-8')
+                else:
+                    # if the article is not in a subdirectory
+                    category = self.settings['DEFAULT_CATEGORY']
 
                 if category != '':
                     metadata['category'] = Category(category, self.settings)
 
-            if 'date' not in metadata and self.settings['DEFAULT_DATE']:
+            if 'date' not in metadata and self.settings.get('DEFAULT_DATE'):
                 if self.settings['DEFAULT_DATE'] == 'fs':
                     metadata['date'] = datetime.datetime.fromtimestamp(
                             os.stat(f).st_ctime)
@@ -299,9 +352,11 @@ class ArticlesGenerator(Generator):
 
             signals.article_generate_context.send(self, metadata=metadata)
             article = Article(content, metadata, settings=self.settings,
-                              filename=f)
+                              filename=f, context=self.context)
             if not is_valid_content(article, f):
                 continue
+
+            self.add_filename(article)
 
             if article.status == "published":
                 if hasattr(article, 'tags'):
@@ -320,7 +375,9 @@ class ArticlesGenerator(Generator):
         for article in self.articles:
             # only main articles are listed in categories, not translations
             self.categories[article.category].append(article)
-            self.authors[article.author].append(article)
+            # ignore blank authors as well as undefined
+            if hasattr(article,'author') and article.author.name != '':
+                self.authors[article.author].append(article)
 
         # sort the articles by date
         self.articles.sort(key=attrgetter('date'), reverse=True)
@@ -396,11 +453,14 @@ class PagesGenerator(Generator):
             except Exception, e:
                 logger.warning(u'Could not process %s\n%s' % (f, str(e)))
                 continue
-            signals.pages_generate_context.send(self, metadata=metadata )
+            signals.pages_generate_context.send(self, metadata=metadata)
             page = Page(content, metadata, settings=self.settings,
-                        filename=f)
+                        filename=f, context=self.context)
             if not is_valid_content(page, f):
                 continue
+
+            self.add_filename(page)
+
             if page.status == "published":
                 all_pages.append(page)
             elif page.status == "hidden":
@@ -436,47 +496,32 @@ class StaticGenerator(Generator):
                  final_path, overwrite=True)
 
     def generate_context(self):
+        self.staticfiles = []
 
-        if self.settings['WEBASSETS']:
-            from webassets import Environment as AssetsEnvironment
-
-            # Define the assets environment that will be passed to the
-            # generators. The StaticGenerator must then be run first to have
-            # the assets in the output_path before generating the templates.
-
-            # Let ASSET_URL honor Pelican's RELATIVE_URLS setting.
-            # Hint for templates:
-            # Current version of webassets seem to remove any relative
-            # paths at the beginning of the URL. So, if RELATIVE_URLS
-            # is on, ASSET_URL will start with 'theme/', regardless if we
-            # set assets_url here to './theme/' or to 'theme/'.
-            # XXX However, this breaks the ASSET_URL if user navigates to
-            # a sub-URL, e.g. if he clicks on a category. To workaround this
-            # issue, I use
-            #     <link rel="stylesheet" href="{{ SITEURL }}/{{ ASSET_URL }}">
-            # instead of
-            #     <link rel="stylesheet" href="{{ ASSET_URL }}">
-            if self.settings.get('RELATIVE_URLS'):
-                assets_url = './theme/'
-            else:
-                assets_url = self.settings['SITEURL'] + '/theme/'
-            assets_src = os.path.join(self.output_path, 'theme')
-            self.assets_env = AssetsEnvironment(assets_src, assets_url)
-
-            if logging.getLevelName(logger.getEffectiveLevel()) == "DEBUG":
-                self.assets_env.debug = True
+        # walk static paths
+        for static_path in self.settings['STATIC_PATHS']:
+            for f in self.get_files(
+                    os.path.join(self.path, static_path), extensions=False):
+                f_rel = os.path.relpath(f, self.path)
+                # TODO remove this hardcoded 'static' subdirectory
+                sc = StaticContent(f_rel, os.path.join('static', f_rel),
+                        settings=self.settings)
+                self.staticfiles.append(sc)
+                self.context['filenames'][f_rel] = sc
+        # same thing for FILES_TO_COPY
+        for src, dest in self.settings['FILES_TO_COPY']:
+            sc = StaticContent(src, dest, settings=self.settings)
+            self.staticfiles.append(sc)
+            self.context['filenames'][src] = sc
 
     def generate_output(self, writer):
-
-        self._copy_paths(self.settings['STATIC_PATHS'], self.path,
-                         'static', self.output_path)
         self._copy_paths(self.settings['THEME_STATIC_PATHS'], self.theme,
                          'theme', self.output_path, '.')
-
-        # copy all the files needed
-        for source, destination in self.settings['FILES_TO_COPY']:
-            copy(source, self.path, self.output_path, destination,
-                 overwrite=True)
+        # copy all StaticContent files
+        for sc in self.staticfiles:
+            mkdir_p(os.path.dirname(sc.save_as))
+            shutil.copy(sc.filepath, sc.save_as)
+            logger.info('copying %s to %s' % (sc.filepath, sc.save_as))
 
 
 class PdfGenerator(Generator):
@@ -519,8 +564,8 @@ class PdfGenerator(Generator):
             try:
                 os.mkdir(pdf_path)
             except OSError:
-                logger.error("Couldn't create the pdf output folder in " + pdf_path)
-                pass
+                logger.error("Couldn't create the pdf output folder in " +
+                             pdf_path)
 
         for article in self.context['articles']:
             self._create_pdf(article, pdf_path)
@@ -541,49 +586,3 @@ class SourceFileGenerator(Generator):
         logger.info(u' Generating source files...')
         for object in chain(self.context['articles'], self.context['pages']):
             self._create_source(object, self.output_path)
-
-class LessCSSGenerator(Generator):
-    """Compile less css files."""
-
-    def _compile(self, less_file, source_dir, dest_dir):
-        base = os.path.relpath(less_file, source_dir)
-        target = os.path.splitext(
-                os.path.join(dest_dir, base))[0] + '.css'
-        target_dir = os.path.dirname(target)
-
-        if not os.path.exists(target_dir):
-            try:
-                os.makedirs(target_dir)
-            except OSError:
-                logger.error("Couldn't create the less css output folder in " +
-                        target_dir)
-
-        subprocess.call([self._lessc, less_file, target])
-        logger.info(u' [ok] compiled %s' % base)
-
-    def generate_output(self, writer=None):
-        logger.info(u' Compiling less css')
-
-        # store out compiler here, so it won't be evaulted on each run of
-        # _compile
-        lg = self.settings['LESS_GENERATOR']
-        self._lessc = lg if isinstance(lg, basestring) else 'lessc'
-
-        # walk static paths
-        for static_path in self.settings['STATIC_PATHS']:
-            for f in self.get_files(
-                    os.path.join(self.path, static_path),
-                    extensions=['less']):
-
-                self._compile(f, self.path, self.output_path)
-
-        # walk theme static paths
-        theme_output_path = os.path.join(self.output_path, 'theme')
-
-        for static_path in self.settings['THEME_STATIC_PATHS']:
-            theme_static_path = os.path.join(self.theme, static_path)
-            for f in self.get_files(
-                    theme_static_path,
-                    extensions=['less']):
-
-                self._compile(f, theme_static_path, theme_output_path)