+++ samples/pelican.conf.py
@@ -12,7 +12,7 @@ LOCALE = "C"
 DEFAULT_PAGINATION = 4
 DEFAULT_DATE = (2012, 03, 02, 14, 01, 01)
 
-FEED_RSS = 'feeds/all.rss.xml'
+FEED_ALL_RSS = 'feeds/all.rss.xml'
 CATEGORY_FEED_RSS = 'feeds/%s.rss.xml'
 
 LINKS = (('Biologeek', 'http://biologeek.org'),
@@ -34,6 +34,9 @@ STATIC_PATHS = ["pictures", ]
 
 # A list of files to copy from the source to the destination
 FILES_TO_COPY = (('extra/robots.txt', 'robots.txt'),)
+
+# custom page generated with a jinja2 template
+TEMPLATE_PAGES = {'pages/jinja2_template.html': 'jinja2_template.html'}
 
 # foobar will not be used, because it's not in caps. All configuration keys
 # have to be in caps